+++ frontend/js/main.js
@@ -1,11 +1,25 @@
+import './helpers';
 import svg4everybody from 'svg4everybody';
 
 /* HMR */
 if (module.hot) module.hot.accept();
 
-window.cl = (...args) => (!PRODUCTION) ? console.log(...args) : false;
-
 document.addEventListener('DOMContentLoaded', () => {
     svg4everybody();
-    cl('Ready');
+
+    /**
+     * async/await
+     * DELETE this code
+     */
+    const mockFetch = () => {
+        return new Promise(resolve => {
+            setTimeout(() => resolve('READY'), 2000);
+        });
+    };
+
+    async function fetchUser() {
+        const user = await mockFetch();
+        cl(user);
+    }
+    fetchUser();
 });