+++ webpack/webpack.config.js
@@ -7,6 +7,7 @@ module.exports = {
     entry: [
         'webpack/hot/dev-server',
         'webpack-hot-middleware/client',
+        'babel-polyfill',        
         path.join(constants.SRC_DIR, 'js/main.js')
     ],
     module: {