+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.0] - 2017-03-04
+### Added
+- eslint + airbnb config
+- babel polyfill -> polyfill Promises
+- babel env preset -> Default behavior without options is the same as babel-preset-latest
+
+### Changed
+
+### Removed
+- latest preset
+
 ## [1.1.0] - 2017-02-25
 ### Added
 - postcss-hexrgba